+++ src/components/BounceActivityContainer/ActivityDetailsModal/index.jsx
