+++ src/components/BounceActivityContainer/ActivityDetailsModal/index.scss
