+++ src/components/BounceActivityContainer/ActivityLogContainer/index.jsx
