+++ src/components/BounceActivityContainer/ActivityLogContainer/index.scss
