+++ src/components/BounceActivityContainer/RuleFilter/index.scss
