+++ src/components/BounceActivityContainer/index.jsx
