+++ src/components/BounceActivityContainer/index.scss
