+++ src/components/BounceActivityContainer/selectors.js
