+++ src/components/BounceDetailsContainer/Changelog/index.scss
