+++ src/components/BounceDetailsContainer/Details/index.jsx
