+++ src/components/BounceDetailsContainer/Modals/ChangeModal/index.scss
