+++ src/components/BounceDetailsContainer/index.scss
